IGAPP-114: Move NavigationTypes to constants
