+++ native/src/modules/app/constants/NavigationTypes.js
