+++ native/src/modules/app/createNavigateToInternalLink.js
@@ -7,8 +7,8 @@ import createNavigateToEvent from './createNavigateToEvent'
 import type { StoreActionType } from './StoreActionType'
 import type { Dispatch } from 'redux'
 import URL from 'url-parse'
-import type { NavigationPropType, RoutesType } from './components/NavigationTypes'
-import { CATEGORIES_ROUTE, DASHBOARD_ROUTE, EVENTS_ROUTE } from './components/NavigationTypes'
+import type { NavigationPropType, RoutesType } from './constants/NavigationTypes'
+import { CATEGORIES_ROUTE, DASHBOARD_ROUTE, EVENTS_ROUTE } from './constants/NavigationTypes'
 import createNavigateToLanding from './createNavigateToLanding'
 import { cityContentPath } from '../common/url'
