+++ native/src/modules/app/navigateToLink.js
@@ -2,8 +2,8 @@
 
 import type { NavigateToInternalLinkParamsType } from './createNavigateToInternalLink'
 import buildConfig from './constants/buildConfig'
-import type { NavigationPropType, RoutesType } from './components/NavigationTypes'
-import { IMAGE_VIEW_MODAL_ROUTE, PDF_VIEW_MODAL_ROUTE } from './components/NavigationTypes'
+import type { NavigationPropType, RoutesType } from './constants/NavigationTypes'
+import { IMAGE_VIEW_MODAL_ROUTE, PDF_VIEW_MODAL_ROUTE } from './constants/NavigationTypes'
 import openExternalUrl from '../common/openExternalUrl'
 
 type NavigateToInternalLinkType = NavigateToInternalLinkParamsType => void
