+++ native/src/modules/endpoint/hocs/withPayloadProvider.js
@@ -11,7 +11,7 @@ import FailureContainer from '../../error/containers/FailureContainer'
 import { LOADING_TIMEOUT } from '../../common/constants'
 import type { ErrorCodeType } from '../../error/ErrorCodes'
 import wrapDisplayName from '../../common/hocs/wrapDisplayName'
-import type { NavigationPropType, RoutePropType, RoutesType } from '../../app/components/NavigationTypes'
+import type { NavigationPropType, RoutePropType, RoutesType } from '../../app/constants/NavigationTypes'
 import { type TFunction } from 'react-i18next'
 import LayoutContainer from '../../layout/containers/LayoutContainer'
 import LayoutedScrollView from '../../common/containers/LayoutedScrollView'
