+++ native/src/routes/events/containers/EventsContainer.js
@@ -18,7 +18,7 @@ import type {
   EventsRouteType,
   NavigationPropType,
   RoutePropType
-} from '../../../modules/app/components/NavigationTypes'
+} from '../../../modules/app/constants/NavigationTypes'
 import navigateToLink from '../../../modules/app/navigateToLink'
 import createNavigateToFeedbackModal from '../../../modules/app/createNavigateToFeedbackModal'
