+++ native/src/routes/offers/containers/OffersContainer.js
@@ -16,12 +16,12 @@ import type {
   OffersRouteType,
   NavigationPropType,
   RoutePropType
-} from '../../../modules/app/components/NavigationTypes'
+} from '../../../modules/app/constants/NavigationTypes'
 import {
   EXTERNAL_OFFER_ROUTE, OFFERS_ROUTE,
   SPRUNGBRETT_OFFER_ROUTE,
   WOHNEN_OFFER_ROUTE
-} from '../../../modules/app/components/NavigationTypes'
+} from '../../../modules/app/constants/NavigationTypes'
 import type { StoreActionType } from '../../../modules/app/StoreActionType'
 import type { Dispatch } from 'redux'
 import LayoutedScrollView from '../../../modules/common/containers/LayoutedScrollView'
