+++ native/src/routes/pdf/components/PDFViewModal.js
@@ -10,7 +10,7 @@ import type {
   PdfViewModalRouteType,
   NavigationPropType,
   RoutePropType
-} from '../../../modules/app/components/NavigationTypes'
+} from '../../../modules/app/constants/NavigationTypes'
 
 type PropsType = {|
   route: RoutePropType<PdfViewModalRouteType>,
